Move integration content into component files for Next.js compatibility
Next.js page files cannot have named exports besides default. Extract
GcpSettingsContent, SmtpSettingsContent, and SlackSettingsContent into
standalone component files under components/settings/. Page files are
now thin wrappers that import the content component.
